Puzzle list fix
